externalize health
