+++ webapp_frontend/src/background/api/registration.ts
@@ -1,0 +1,7 @@
+
+
+function test(){
+    return null
+}
+
+export default test;
