+++ webapp_frontend/src/components/basicElements/Footer.tsx
@@ -1,7 +1,11 @@
 import React, {ReactElement} from 'react';
 
 function Footers():ReactElement {
-    return(<div>hallo</div>);
+    return(
+        <div>
+            hallo
+        </div>
+    );
 }
 
 export default Footers;
